test: move live probes under test/live/
Issue #151 remaining physical isolation. Production probes now live
under test/live/, npm test uses a non-recursive glob so those files
cannot run inside the deterministic suite, and npm run test:live
selects only that directory. Shared JSON-schema subset extracted so
the live schema file does not import a .test.ts module.
