+++ test/live-probe-gate.test.ts
@@ -80,6 +80,9 @@ test("every test that reads the deployment is behind the live-probe gate", () =>
 test("the deterministic suite is the default and the live one is opt-in", () => {
   const pkg = JSON.parse(readFileSync(new URL("../package.json", import.meta.url), "utf8"));
   assert.ok(!/LIVE_PROBES/.test(pkg.scripts.test), "`npm test` must not turn the probes on");
+  assert.ok(!pkg.scripts.test.includes("**"), "`npm test` must not recurse into test/live/");
+  assert.ok(pkg.scripts.test.includes("test/*.test.ts"));
+  assert.ok(pkg.scripts["test:live"].includes("test/live/*.test.ts"));
   assert.match(pkg.scripts["test:live"], /LIVE_PROBES=1/, "`npm run test:live` turns them on");
   assert.ok(pkg.scripts["test:all"], "and there is one command that runs both");
 });
@@ -89,7 +92,7 @@ test("the live lane does not skip on unreachability or a missing deployment mark
   // live run could mean "could not check". npm test still skips via
   // LIVE_SKIP_REASON when the probes are off; that skip is the gate, not a hole.
   const dir = new URL("./", import.meta.url);
-  for (const f of ["schema.test.ts", "param-home.test.ts"]) {
+  for (const f of ["live/schema.test.ts", "live/param-home.test.ts"]) {
     const src = readFileSync(new URL(f, dir), "utf8");
     assert.equal(
       /t\.skip\(`API unreachable/.test(src),
@@ -153,7 +156,7 @@ test("the live-probe inventory is the three files that call liveFetch", () => {
   // #151 remaining: a mechanical inventory of production probes. Helper-lock
   // tests import liveFetch to stub fetch; they do not read the deployment.
   // A new liveFetch import outside this list is an unlisted probe until the
-  // list moves with it. The physical move under test/live/ is a later slice.
+  // list moves with it. The physical move under test/live/ is a this slice.
   const dir = new URL("./", import.meta.url);
   const callers: string[] = [];
   for (const f of testFiles(dir)) {
@@ -161,12 +164,12 @@ test("the live-probe inventory is the three files that call liveFetch", () => {
     if (f === "live-fetch-lock.test.ts") continue;
     if (f === "live-probe-gate.test.ts") continue;
     const src = readFileSync(new URL(f, dir), "utf8");
-    if (!/from "\.\/helpers\/live\.ts"/.test(src)) continue;
+    if (!/helpers\/live\.ts/.test(src)) continue;
     if (/(?<![.\w])liveFetch\s*\(/.test(src)) callers.push(f);
   }
   assert.deepEqual(
     callers.sort(),
-    ["ledger-tx-migration.test.ts", "param-home.test.ts", "schema.test.ts"],
+    ["live/ledger-tx-migration.test.ts", "live/param-home.test.ts", "live/schema.test.ts"],
     `liveFetch callers changed: ${callers.join(", ")}`,
   );
 });
